requirements.txt
pandas==1.2.4
psycopg2-binary==2.8.6
